Dobavil interfeys, rabotu s konfiguratsionnym faylom 51/5000 Added an interface, working with a configuration file
Added interface:
Now, when the script starts, a window appears with input fields and 2 buttons and 3 checkboxes. If a configuration file exists in the folder, then the last configuration is read from. Only floating point numbers can be entered in the input field. Jackdaws define flags. The "Save config" button saves the configuration file to the script folder.
The "Generate gcode" button generates gcode using the entered values ​​in the fields, leaving the configuration file unchanged.
To work with the configuration file, it will create a class that stores and processes all actions:
writing to the class of values ​​from input fields (updatesettings);
Method for interacting with the button (update_and_create);
The method for saving the configuration file (save_config);
Method for reading the configuration file (read_config).

The logic of the script itself remained the same, only it was modernized for working with the class.